+++ src/components/common/index/style.css
@@ -169,7 +169,7 @@ button:active {
 	height:40px;
 	padding-left:8px;
 	font-size:14px;
-	background:#FF0543;
+	background:#0087d0;
 	text-align:left;
 	line-height:40px;
 }
